Fixed several bugs in image build process (#304)
* added additional logging to build manager

* bug fix: bad blueprint construction

* bug fix: only destroy .Dockerfile if it exists

* added context var

* bug fix: infinite loop

* updated CHANGELOG

* increased version number

+++ bugzoo/mgr/build.py
@@ -2,11 +2,15 @@ from typing import Iterator
 import os
 import shutil
 import json
+import logging
 
 import docker
 
 from ..core.build import BuildInstructions
 from ..exceptions import ImageBuildFailed
+
+logger = logging.getLogger(__name__)  # type: logging.Logger
+logger.setLevel(logging.DEBUG)
 
 
 class BuildManager(object):
@@ -86,23 +90,26 @@ class BuildManager(object):
             quiet: used to enable and disable output from the Docker build
                 process.
         """
+        logger.debug("request to build image: %s", name)
         instructions = self[name]
 
         if instructions.depends_on:
+            logger.info("building dependent image: %s",
+                        instructions.depends_on)
             self.build(instructions.depends_on, force=force, quiet=quiet)
 
         if not force and self.is_installed(instructions.name):
             return
 
-        # TODO use logger
         if not quiet:
-            print("Building image: {}".format(name))
+            logger.info("building image: %s", name)
 
-        tf = os.path.join(instructions.abs_context, '.Dockerfile')
+        context = instructions.abs_context
+        tf = os.path.join(context, '.Dockerfile')
         try:
             success = False
-            shutil.copy(instructions.file_abs, tf)
-            response = self.__docker.api.build(path=instructions.abs_context,
+            shutil.copy(instructions.filename_abs, tf)
+            response = self.__docker.api.build(path=context,
                                                dockerfile='.Dockerfile',
                                                tag=name,
                                                # pull=force,
@@ -124,10 +131,11 @@ class BuildManager(object):
                 raise ImageBuildFailed(name, log)
 
             if success and not quiet:
-                print("Built image: {}".format(name))
+                logger.info("built image: %s", name)
                 return
         finally:
-            os.remove(tf)
+            if os.path.exists(tf):
+                os.remove(tf)
 
     def uninstall(self,
                   name: str,
